add common.js for some constats and methods
