+++ src/index.js
@@ -1,9 +1,6 @@
 import readlinesync from 'readline-sync';
-
-const roundsGame = 3;
-const maxInt = 100;
-
-const getRandomInt = () => Math.floor(Math.random() * maxInt);
+import greetings from './cli.js';
+import { getRandomInt, testEven, roundsGame } from './common.js';
 
 const wrongAnswerShowText = (userName, rightAnswer, answerUser) => {
   console.log(`'${answerUser}' is wrong answer ;(. Correct answer was '${rightAnswer}'.`);
@@ -28,6 +25,28 @@ const calculateUserQuestions = (userName, rightAnswer, correctAnswers) => {
   return [countRightAnswers, answerUser];
 };
 
-export {
-  getRandomInt, calculateUserQuestions, congratulatUser, roundsGame,
+const getRightAnswer = (gameName) => {
+  if (gameName === 'Even') {
+    console.log('Answer "yes" if the number is even, otherwise answer "no".');
+    const randNum = getRandomInt();
+    console.log(`Question: ${randNum}`);
+    return testEven(randNum);
+  }
+  return 1;
 };
+
+const engineGames = (gameName) => {
+  const userName = greetings();
+  let countRound = 0;
+  let correctAnswers = 0;
+  let answerUser;
+  let rightAnswer;
+  do {
+    countRound += 1;
+    rightAnswer = getRightAnswer(gameName);
+    [correctAnswers, answerUser] = calculateUserQuestions(userName, rightAnswer, correctAnswers);
+  } while ((countRound < roundsGame) && (answerUser === rightAnswer));
+  congratulatUser(userName, correctAnswers);
+};
+
+export default engineGames;
